Move network to separate component
